Updated README.md
Fixed a misspell.
"Notice an user/channel" to "Notice a user/channel"

Super detailed explaination:
'an' is the same as 'a'. The only difference is that "an" is used before a word that starts with the sound of a vowel. "user" starts with a vowel, but it does not start with the sound of a vowel.